SQL project
Full description of this database could be found in the pdf file. This is the database for the vegetable warehouse. You can find the file of creation db, file of filling the db, ER diagram, and files with some triggers, selections, procedures, functions and etc. The most interesting part is about creating orders and automatic filling tables.